Successfull build after upgrade

+++ openmeetings-core/src/main/java/org/apache/openmeetings/core/converter/ImageConverter.java
@@ -87,20 +87,28 @@ public class ImageConverter extends BaseConverter {
 		return logs;
 	}
 
-	public ProcessResultList convertImageUserProfile(File file, Long userId) throws Exception {
+	public ProcessResultList convertImageUserProfile(File file, Long userId, boolean skipConvertion) throws Exception {
 		ProcessResultList returnMap = new ProcessResultList();
 
 		// User Profile Update
-		Files.newDirectoryStream(
-				getUploadProfilesUserDir(userId).toPath()
-				, fi -> fi.toString().endsWith(EXTENSION_PNG))
-					.forEach(path -> FileUtils.deleteQuietly(path.toFile()));
+		File[] files = getUploadProfilesUserDir(userId).listFiles(fi -> fi.getName().endsWith(EXTENSION_PNG));
+		if (files != null) {
+			for (File f : files) {
+				FileUtils.deleteQuietly(f);
+			}
+		}
 
 		File destinationFile = OmFileHelper.getNewFile(getUploadProfilesUserDir(userId), PROFILE_FILE_NAME, EXTENSION_PNG);
-		returnMap.add(resize(file, destinationFile, 250, 250, true));
+		if (!skipConvertion) {
+			returnMap.add(convertSinglePng(file, destinationFile));
+		} else {
+			FileUtils.copyFile(file, destinationFile);
+		}
 
-		// Delete old one
-		Files.deleteIfExists(file.toPath());
+		if (!skipConvertion) {
+			// Delete old one
+			Files.deleteIfExists(file.toPath());
+		}
 
 		String img = destinationFile.getName();
 		User us = userDao.get(userId);
@@ -151,9 +159,9 @@ public class ImageConverter extends BaseConverter {
 		return ProcessHelper.exec("convertSinglePng", argv);
 	}
 
-	public ProcessResult resize(File in, File out, Integer width, Integer height, boolean max) throws IOException {
+	public ProcessResult resize(File in, File out, Integer width, Integer height) throws IOException {
 		List<String> argv = List.of(getPathToConvert()
-				, "-resize", (width == null ? "" : width) + (height == null ? "" : "x" + height) + (max ? ">" : "")
+				, "-resize", (width == null ? "" : width) + (height == null ? "" : "x" + height)
 				, in.getCanonicalPath(), out.getCanonicalPath());
 		return ProcessHelper.exec("resize", argv);
 	}
